requirements.txt
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
kiwisolver==1.4.5
matplotlib==3.8.4
metalcompute==0.2.5
numpy==1.26.3
packaging==24.0
pillow==10.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
